requirements.txt
argcomplete ~= 1.12
colorlog ~= 6.6
GitPython ~= 3.1
ordered-set ~= 4.0
readchar ~= 3.0
version-query ~= 1.1
